Split up query mod
