+++ src/lib.rs
@@ -16,14 +16,17 @@ mod archetype;
 mod archetype_set;
 mod borrow_checker;
 mod column;
-mod query;
+
+pub mod query;
 
 pub use thunderdome;
 
 pub use archetype::{Archetype, Entity, EntityColumns, EntityKey};
 pub use archetype_set::{ArchetypeSet, ArchetypeSetFetch, EntityId, InArchetypeSet};
 pub use column::Column;
-pub use query::{Fetch, Query};
+
+#[doc(inline)]
+pub use query::Query;
 
 pub trait Component: 'static {}
 
@@ -31,5 +34,5 @@ impl<T> Component for T where T: 'static {}
 
 // Hidden unstable symbols, needed for `stecs-derive`.
 pub mod internal {
-    pub use super::{borrow_checker::BorrowChecker, query::FetchEntityId};
+    pub use super::borrow_checker::BorrowChecker;
 }
