--- src/query/stream.rs
@@ -1,3 +1,0 @@
-pub struct Stream {
-    
-}
